+++ changelog.md
@@ -1,5 +1,9 @@
-0.6.0
-PascalCase all schema Collection Names
-0.7.0
+## 0.8.0
+use multiple and required in FileField to determine either string string[] or undefined
+take required into account for all fields that have it
+## 0.7.0
 Complete rewrite, replace json-to-ts with custom interface creation logic
+## 0.6.0
+PascalCase all schema Collection Names
+